The SettingsApp has been routed under /organizations/:organizationId/settings for now ... Will review later how to have it shifted to specific settings route like /users/settings This has been done to tackle the current Organisation set in redux store currently which affects the TopBar component as well ... This is temporary route as of now & will be revised once the whole component is working fine Currently efforts in progress to put conversation members actions and reducers in place required for settings screen

+++ app/react/components/ProfileDropdown/ProfileDropdown.jsx
@@ -6,30 +6,32 @@ class ProfileDropdown extends React.Component {
     currentUser: PropTypes.shape({
       name: PropTypes.string.isRequired,
     }).isRequired,
+    currentOrganization: PropTypes.object
   };
 
   static defaultProps = {};
 
   renderOpen = () => {
     const { name } = this.props.currentUser;
-
-    return (
-      <div className={styles.rootOpen}>
-        <div className={styles.header}>
-          <div className={styles.round} title={name} >
-            {name.charAt(0)}
+    if (this.props.currentOrganization) {
+      return (
+          <div className={styles.rootOpen}>
+            <div className={styles.header}>
+              <div className={styles.round} title={name}>
+                {name.charAt(0)}
+              </div>
+              <div className={styles.name}>
+                {name}
+              </div>
+              <div className={styles.caretOpen}></div>
+            </div>
+            <div className={styles.menu}>
+              <a href={"/organizations/" + this.props.currentOrganization.id + "/settings"}>Settings</a>
+              <a href="/users/sign_out" data-method="delete">Sign Out</a>
+            </div>
           </div>
-          <div className={styles.name}>
-            {name}
-          </div>
-          <div className={styles.caretOpen}></div>
-        </div>
-        <div className={styles.menu}>
-          <a href="/users/settings">Settings</a>
-          <a href="/users/sign_out" data-method="delete">Sign Out</a>
-        </div>
-      </div>
-    );
+      );
+    }
   };
 
   render() {
